% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R, R/x13_spec.R
\name{jd3_utilities}
\alias{jd3_utilities}
\alias{jd2r_spec_x11}
\alias{r2jd_spec_x11}
\alias{r2jd_spec_regarima}
\alias{jd2r_spec_regarima}
\alias{r2jd_spec_x13}
\alias{jd2r_spec_x13}
\title{Java Utility Functions}
\usage{
jd2r_spec_x11(spec)

r2jd_spec_x11(spec)

r2jd_spec_regarima(spec)

jd2r_spec_regarima(spec)

r2jd_spec_x13(spec)

jd2r_spec_x13(spec)
}
\arguments{
\item{jtest, s, period, startYear, startPeriod, length, ldt, dt, jparams, start, end, y, m, d, year, month, day, type, p, code, prefix, r, span, rspan, rslt, name, jobj, jrslt}{parameters.}
}
\description{
These functions are used in all JDemetra+ 3.0 packages to easily interact between R and Java objects.
}
