% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regarima_outliers.R
\name{regarima.outliers}
\alias{regarima.outliers}
\title{Title}
\usage{
regarima.outliers(
  y,
  order = c(0L, 1L, 1L),
  seasonal = c(0L, 1L, 1L),
  mean = F,
  X = NULL,
  X.td = NULL,
  ao = T,
  ls = T,
  tc = F,
  so = F,
  cv = 0
)
}
\arguments{
\item{cv}{}
}
\value{

}
\description{
Title
}
