% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regarima_outliers.R
\name{regarima_outliers}
\alias{regarima_outliers}
\title{Detect Outliers in RegARIMA Model}
\usage{
regarima_outliers(
  y,
  order = c(0L, 1L, 1L),
  seasonal = c(0L, 1L, 1L),
  mean = F,
  X = NULL,
  X.td = NULL,
  ao = T,
  ls = T,
  tc = F,
  so = F,
  cv = 0
)
}
\arguments{
\item{y}{the dependent variable (\code{ts} object).}

\item{order, seasonal}{the orders of the ARIMA model.}

\item{mean}{boolean to include or not the mean.}

\item{X}{explanatory varibales.}

\item{X.td}{trading days regressors.}

\item{ao, ls, so, tc}{boolean to indicate which outliers are detected.}

\item{cv}{\code{numeric}. The entered critical value for the outliers' detection procedure.
If equal to 0 the critical value for the outliers' detection procedure is automatically determined
by the number of observations.}
}
\value{
a \code{"JDSTS"} object.
}
\description{
Detect Outliers in RegARIMA Model
}
\examples{
regarima_outliers(rjd3toolkit::ABS$X0.2.09.10.M)
}
