% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x13_spec.R
\name{spec_regarima_default}
\alias{spec_regarima_default}
\title{RegARIMA Default Specification}
\usage{
spec_regarima_default(name = "rg4")
}
\arguments{
\item{name}{the name of a predifined specification (`"rg0"`, `"rg1"`, `"rg2c"`, `"rg3"`, `"rg4"` or `"rg5c"`, see details).#'}
}
\value{

}
\description{
RegARIMA Default Specification
}
