% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udvar.R
\name{userdefined_variables_x13}
\alias{userdefined_variables_x13}
\title{Retrieve the user-defined variable names}
\usage{
userdefined_variables_x13(x = c("X-13", "RegArima", "X-11"))
}
\arguments{
\item{x}{a character}
}
\description{
Function to retrieve the names of the additional output variables that can be defined in \code{x13()}, \code{regarima()} and \code{x11()} functions.
}
\examples{
userdefined_variables_x13("x13")
userdefined_variables_x13("regarima")
userdefined_variables_x13("x11")
}
