% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x13.R
\name{x11}
\alias{x11}
\title{X11 Decomposition Algorithm}
\usage{
x11(ts, spec = spec_x11_default())
}
\arguments{
\item{ts}{a univariate time series}

\item{spec}{the specification}
}
\value{

}
\description{
X11 Decomposition Algorithm
}
