% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x13_spec.R
\name{x13_spec}
\alias{x13_spec}
\alias{spec_regarima_default}
\alias{spec_x13_default}
\alias{spec_x11_default}
\title{RegARIMA/X-13 Default Specification}
\usage{
spec_regarima_default(name = c("rg4", "rg0", "rg1", "rg2c", "rg3", "rg5c"))

spec_x13_default(name = c("rsa4", "rsa0", "rsa1", "rsa2c", "rsa3", "rsa5c"))

spec_x11_default()
}
\arguments{
\item{name}{the name of a predefined specification.}
}
\value{
an object of class \code{"JD3_X13_SPEC"} (\code{spec_x13_default()}),
\code{"JD3_REGARIMA_SPEC"} (\code{spec_regarima_default()}) or
\code{"JD3_X11_SPEC"} (\code{spec_x11_default()}).
}
\description{
Set of functions to create default specification associated to the X-13ARIMA seasonal adjustment method
}
\details{
The available predefined 'JDemetra+' model specifications are described in the table below:

\tabular{rrrrrrr}{
\strong{Identifier} |\tab \strong{Log/level detection} |\tab \strong{Outliers detection} |\tab \strong{Calendar effects} |\tab \strong{ARIMA}\cr
RSA0/RG0 |\tab \emph{NA} |\tab \emph{NA} |\tab \emph{NA} |\tab Airline(+mean)\cr
RSA1/RG1 |\tab automatic |\tab AO/LS/TC  |\tab \emph{NA} |\tab Airline(+mean)\cr
RSA2c/RG2c |\tab automatic |\tab AO/LS/TC |\tab 2 td vars + Easter |\tab Airline(+mean)\cr
RSA3/RG3 |\tab automatic |\tab AO/LS/TC |\tab \emph{NA} |\tab automatic\cr
RSA4c/RG4c |\tab automatic |\tab AO/LS/TC |\tab 2 td vars + Easter |\tab automatic\cr
RSA5c/RG5c |\tab automatic |\tab AO/LS/TC |\tab 7 td vars + Easter |\tab automatic
}
}
