% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udvar.R
\name{userdefined_variables_x13}
\alias{userdefined_variables_x13}
\title{List of all available estimation output objects (series, parameters, diagnostics)}
\usage{
userdefined_variables_x13(x = c("X-13", "RegArima", "X-11"))
}
\arguments{
\item{x}{a character to indicate the estimation function for which the output items list will be edited.}
}
\value{
a vector of string conataing all the available output objects (series, diagnostics, parameters)
}
\description{
Function generating a comprehensive list of available output variables generated in the estimation process
by the  \code{x13()}, \code{regarima()} and \code{x11()} functions.
Some items are available in the default estimation output but the remainder can be added
using the \code{userdefined} parameter
User-defined objects can the be retrieved from the list of lists generated by the estimation process
}
\examples{
userdefined_variables_x13("x13")
userdefined_variables_x13("regarima")
userdefined_variables_x13("x11")
y = rjd3toolkit::ABS$X0.2.09.10.M
}
