% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x13.R
\name{x13}
\alias{x13}
\alias{fast.x13}
\title{Seasonal Adjustment with  X13-ARIMA}
\usage{
x13(ts, spec = "rsa4", context = NULL, userdefined = NULL)

fast.x13(ts, spec = "rsa4", context = NULL, userdefined = NULL)
}
\arguments{
\item{ts}{a univariate time series.}

\item{spec}{the model specification. Can be either the name of a predefined specification or a user-defined specification.}

\item{context}{the dictionnary of variables.}

\item{userdefined}{a vector containing the additional output variables.}
}
\description{
Seasonal Adjustment with  X13-ARIMA
}
\examples{
sp = spec_x13_default("rg5c")
y = rjd3toolkit::ABS$X0.2.09.10.M
fast.x13(y, spec = sp)
sp = add_outlier(sp,
                 type = c("AO"), c("2015-01-01", "2010-01-01"))
sp = set_transform(
  set_tradingdays(
    set_easter(sp, enabled = FALSE),
    option = "workingdays"
  ),
  fun = "None"
)
sp = set_x11(sp,
             henderson.filter = 13)
fast.x13(y, spec = sp)

}
